unintuitive, but fixed
argparse's 'store_true' and 'store_false' indicate that those are the
actions to perform _if the flag is present_, otherwise, the opposite
will be stored. so if you set a flag like --myflag and put
action=store_true but then don't pass that flag into your command-line
program, then False will automatically be stored.

because of this the "default" value (e.g., when the flag is not passed
into the program) will usually be the opposite of what's listed in the
action='store_*'